feat: added prebuild script for all environments
