+++ package.json
@@ -13,6 +13,7 @@
     "test": "jest --roots src --watch",
     "test:ci": "jest --roots src",
     "dev": "yarn docs:build && next dev",
+    "prebuild": "node prebuild.js",
     "build": "yarn docs:build && next build",
     "start": "next start",
     "pretty": "prettier -w .",
